requirements.txt

Flask==2.2.5
gunicorn==20.1.0
numpy==1.24.3
Pillow==10.1.0
tensorflow-cpu==2.12.0
flask-cors==3.0.10

